Event aside shows next event when available
When there is an upcoming event, dynamically show those details in the aside. If
not, the next event date is shown, which is the next third Thursday of the
month.

Note that there is now a setup/teardown on all tests to jump to a fixed known
date. This makes it easier to write tests and fixtures with known good dates.