Add CSS and JS for hints in content
These will be used in quizzes to provide a mechanism to provide initially-hidden
hints on how to solve a given problem. The markup should look something like
this:

`
<div class="hint">
	<a href="#" class="hint-reveal">Click to reveal a hint to help you with Stage 1.</a>
	<p>
		<strong>Hint:</strong> Hint text here.
	</p>
</div>
`

This behaves in a progressive manner - without JavaScript, the hint
text will be visible and the link to reveal will not be displayed.