Bump Ruby to 3.3.6
This gets us to a more modern Ruby. Not going all the way to 3.4 just
yet, will do that once we're on Rails 7, as that includes a bunch of
default gems that get removed in 3.4 (and I can't be bothered to add
them to the Gemspec only to remove them again later).